+++ changelog.md
@@ -11,6 +11,7 @@
 - **Changed:**
   - Update electron 9.1.0 to 10.1.1
   - Update chokidar 3.4.0 to 3.4.2
+  - Update yargs-parser 18.1.3 to 20.0.0
 
 ## v0.12.2 2020-07-10